PAGE RE-RENDER ERROR FIXED:
Created a useEffect and called the checkLogin() funtion within it, setting the useEffect to only load/render once.
Set the Login state to null by default and the setting the login state to the current logged in user after user logs in.
Removed the getElement by ID function previouly used for displaying the appropriate user based on the login State.
Tested the update feature with different browers with no error, Error Solved.